requirements.txt
streamlit
transformers>=4.40.0
sentence-transformers
faiss-cpu
pandas
torch
